requirements.txt
annotated-types==0.7.0
anthropic==0.50.0
anyio==4.9.0
arq==0.25.0
backoff==2.2.1
beautifulsoup4==4.13.4
boto3==1.38.1
botocore==1.38.1
browser-use @ git+https://github.com/browser-use/browser-use.git@5674796a2bdd9170eb30ccfe2fdf174bd51238a1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
faiss-cpu==1.10.0
fastapi==0.111.1
fastapi-cli==0.0.7
filetype==1.2.0
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.39.0
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h2==4.2.0
hiredis==3.1.0
hpack==4.1.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.22
langchain-anthropic==0.3.3
langchain-aws==0.2.19
langchain-core==0.3.49
langchain-deepseek==0.1.3
langchain-google-genai==2.1.2
langchain-ollama==0.3.0
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langsmith==0.3.33
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==3.0.2
mdurl==0.1.2
mem0ai==0.1.93
monotonic==1.6
numpy==1.26.4
ollama==0.4.8
openai==1.76.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pandas-stubs==2.2.3.250308
patchright==1.51.3
playwright==1.51.0
portalocker==2.10.1
posthog==3.25.0
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic-settings==2.2.1
pydantic_core==2.27.2
pyee==12.1.1
Pygments==2.19.1
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2024.2
pywin32==310
PyYAML==6.0.2
qdrant-client==1.14.1
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.3
rsa==4.9.1
s3transfer==0.12.0
screeninfo==0.8.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
starlette==0.37.2
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typer==0.15.2
types-pytz==2025.2.0.20250326
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.29.0
watchfiles==1.0.5
websockets==15.0.1
zstandard==0.23.0
